+++ CHANGELOG.md
@@ -28,6 +28,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### New features
 
+* Support background filesystem monitoring via watchman triggers enabled with
+  the `core.watchman.register_trigger = true` config.
+
 * Show paths to config files when configuration errors occur
 
 * `jj fix` now supports configuring the default revset for `-s` using the